+++ Tests/ChessKitEngineTests/EngineTests/Lc0Tests.swift
@@ -3,8 +3,8 @@
 //  ChessKitEngineTests
 //
 
-import XCTest
 @testable import ChessKitEngine
+import XCTest
 
 final class Lc0Tests: BaseEngineTests {
 
@@ -13,31 +13,4 @@ final class Lc0Tests: BaseEngineTests {
     super.setUp()
   }
 
-  override func testEngineRestart() async {
-    XCTAssert(!Thread.isMainThread, "Test must be run on a background thread")
-    XCTAssertNotNil(self.engine, "Failed to initialize engine")
-
-    let expectationStartEngine = self.expectation(
-      description: "Expect engine \(engine.type.name) to start up."
-    )
-    let expectationStopEngine = self.expectation(
-      description: "Expect engine \(engine.type.name) to stop gracefully."
-    )
-
-    expectationStartEngine.expectedFulfillmentCount = 2
-
-    await startEngine(expectation: expectationStartEngine)
-    await stopEngine(expectation: expectationStopEngine)
-
-    // lc0 has an internal mutex failure "Unhandled exception: mutex lock failed: Invalid argument"
-    // when trying to stop and start the engine too fast.
-    // This 100 ms delay circumvents the issue.
-    //
-    // Once this issue is resolved, this function
-    // can be removed, using `BaseEngineTests.testEngineRestart()` instead.
-    try? await Task.sleep(for: .milliseconds(100))
-    await startEngine(expectation: expectationStartEngine)
-
-    await fulfillment(of: [expectationStartEngine, expectationStopEngine], timeout: 5)
-  }
 }