+++ Tests/ChessKitEngineTests/EngineTests/StockfishTests.swift
@@ -3,16 +3,14 @@
 //  ChessKitEngineTests
 //
 
-//import XCTest
-//@testable import ChessKitEngine
-//
-//final class StockfishTests: BaseEngineTests {
-//
-//    override func setUp() {
-//        engineType = .stockfish
-//        super.setUp()
-//    }
-//
-//}
+@testable import ChessKitEngine
+import XCTest
 
-// ↑↑ Failing in CI since update to Stockfish 17 - to be investigated. ↑↑
+final class StockfishTests: BaseEngineTests {
+
+  override func setUp() {
+    engineType = .stockfish
+    super.setUp()
+  }
+
+}