+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ### Improvements
 * Commands and responses are now logged using their own synchronous queue to prevent unreadable output when commands and responses were sent simultaneously.
-* `setLoggingEnabled(loggingEnabled:)` has been renamed `set(loggingEnabled:)`
+* `setLoggingEnabled(loggingEnabled:)` has been renamed `set(loggingEnabled:) async`
   * The original function remains but is deprecated.
+* Change default `coreCount` in `Engine.start()` to **one less** than the core count (minimum of 1).
+  * This allows a free core for other processing tasks in a multicore environment.
+  * Can still be changed with the `coreCount` parameter if desired.
 
 # ChessKitEngine 0.6.0
 Released Friday, May 30, 2025.